⭐️ expose windows wsus updates as `os.updates`
