+++ resources/packs/core/packages/windows_packages_test.go
@@ -4,9 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/cockroachdb/errors"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"go.mondoo.com/cnquery/motor/providers/mock"
 	"go.mondoo.com/cnquery/motor/providers/os/powershell"
 )
@@ -72,46 +70,4 @@ func TestWindowsHotFixParser(t *testing.T) {
 	hotfixes, err = ParseWindowsHotfixes(strings.NewReader(""))
 	assert.Nil(t, err)
 	assert.Equal(t, 0, len(hotfixes), "detected the right amount of packages")
-}
-
-func TestWinOSUpdatesParser(t *testing.T) {
-	mock, err := mock.NewFromTomlFile("./testdata/windows_2019.toml")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cmd := powershell.Encode(WINDOWS_QUERY_WSUS_AVAILABLE)
-	c, err := mock.RunCommand(cmd)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	m, err := ParseWindowsUpdates(c.Stdout)
-	assert.Nil(t, err)
-	assert.Equal(t, 6, len(m), "detected the right amount of packages")
-
-	pkg, err := findKb(m, "890830")
-	require.NoError(t, err)
-	assert.Equal(t, "890830", pkg.Name, "update id detected")
-	assert.Equal(t, "Windows Malicious Software Removal Tool x64 - March 2020 (KB890830)", pkg.Description, "update title detected")
-
-	pkg, err = findKb(m, "4538461")
-	require.NoError(t, err)
-	assert.Equal(t, "4538461", pkg.Name, "update id detected")
-	assert.Equal(t, "2020-03 Cumulative Update for Windows Server 2019 (1809) for x64-based Systems (KB4538461)", pkg.Description, "update title detected")
-
-	// check empty return
-	m, err = ParseWindowsUpdates(strings.NewReader(""))
-	assert.Nil(t, err)
-	assert.Equal(t, 0, len(m), "detected the right amount of packages")
-}
-
-func findKb(pkgs []Package, name string) (Package, error) {
-	for i := range pkgs {
-		if pkgs[i].Name == name {
-			return pkgs[i], nil
-		}
-	}
-
-	return Package{}, errors.New("not found")
 }
